Conflicts:
	recipes/agent_registration.rb

+++ providers/source.rb
@@ -1,11 +1,11 @@
 action :extract_only do
   tar_path = zabbix_tar_path(new_resource.code_dir, new_resource.branch, new_resource.version)
 
-  unless ::File.exists?(tar_path)
+  if !::File.exist?(tar_path)
     Chef::Log.info("Zabbix tar: #{tar_path} does't exist")
     remote_file tar_path do
       source new_resource.source_url
-      mode "0644"
+      mode '0644'
       action :create
     end
     new_resource.updated_by_last_action(true)
@@ -14,19 +14,19 @@ action :extract_only do
   end
 
   extract_to = extract_dir(new_resource.code_dir, new_resource.target_dir)
-  tmp_dir = ::File.join("/tmp", "zabbix-#{new_resource.version}")
-  unless ::File.exists?(extract_to)
+  tmp_dir = ::File.join('/tmp', "zabbix-#{new_resource.version}")
+  if !::File.exist?(extract_to)
     Chef::Log.info("Zabbix extract: #{extract_to} doesn't exist")
     script "extract Zabbix to #{extract_to}" do
-      interpreter "bash"
-      user "root"
+      interpreter 'bash'
+      user 'root'
       code <<-EOH
         rm -rf #{tmp_dir}
         tar xvfz #{tar_path} -C /tmp
         mv #{tmp_dir} #{extract_to}
       EOH
 
-      not_if { ::File.exists?(extract_to) }
+      not_if { ::File.exist?(extract_to) }
     end
     new_resource.updated_by_last_action(true)
   else
@@ -38,11 +38,11 @@ action :install_server do
   action_extract_only
 
   source_dir = extract_dir(new_resource.code_dir, new_resource.target_dir)
-  unless ::File.exists?(::File.join(source_dir, "already_built")) 
+  unless ::File.exist?(::File.join(source_dir, 'already_built'))
     Chef::Log.info("Compiling Zabbix Server with options '#{new_resource.configure_options}")
     script "install_zabbix_server_#{zabbix_source_identifier(new_resource.branch, new_resource.version)}" do
-      interpreter "bash"
-      user "root"
+      interpreter 'bash'
+      user 'root'
       code <<-EOH
         (cd #{source_dir} && ./configure --enable-server --prefix=#{new_resource.install_dir} #{new_resource.configure_options})
         (cd #{source_dir} && make install && touch already_built)
@@ -56,11 +56,11 @@ action :install_agent do
   action_extract_only
 
   source_dir = extract_dir(new_resource.code_dir, new_resource.target_dir)
-  unless ::File.exists?(::File.join(source_dir, "already_built")) 
+  unless ::File.exist?(::File.join(source_dir, 'already_built'))
     Chef::Log.info("Compiling Zabbix Agent with options '#{new_resource.configure_options}")
     script "install_zabbix_agent_#{zabbix_source_identifier(new_resource.branch, new_resource.version)}" do
-      interpreter "bash"
-      user "root"
+      interpreter 'bash'
+      user 'root'
       code <<-EOH
       (cd #{source_dir} && ./configure --enable-agent --prefix=#{new_resource.install_dir} #{new_resource.configure_options})
       (cd #{source_dir} && make install && touch already_built)
@@ -71,12 +71,12 @@ action :install_agent do
 end
 
 def load_current_resource
-  run_context.include_recipe "zabbix::_providers_common"
+  run_context.include_recipe 'zabbix::_providers_common'
   require 'zabbixapi'
 end
 
 def zabbix_source_identifier(branch, version)
-  "#{branch.gsub("%20", "-")}-#{version}"
+  "#{branch.gsub('%20', '-')}-#{version}"
 end
 
 def zabbix_tar_path(code_dir, branch, version)
